scripts/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
python-dateutil
lxml 